Make STI scripts downloading and git clone executed in parallel
