+++ sti/util.go
@@ -2,9 +2,15 @@ package sti
 
 import (
 	"archive/tar"
+	"bufio"
 	"bytes"
+	"encoding/json"
+	"fmt"
 	"io"
 	"io/ioutil"
+	"log"
+	"net/http"
+	"net/url"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -12,6 +18,37 @@ import (
 
 	"github.com/fsouza/go-dockerclient"
 )
+
+// SchemeReaders create an io.Reader from the given url.
+//
+type SchemeReader func(*url.URL) (io.ReadCloser, error)
+
+var schemeReaders = map[string]SchemeReader{
+	"http":  readerFromHttpUrl,
+	"https": readerFromHttpUrl,
+	"file":  readerFromFileUrl,
+}
+
+// This SchemeReader can produce an io.Reader from a file URL.
+//
+func readerFromHttpUrl(url *url.URL) (io.ReadCloser, error) {
+	resp, err := http.Get(url.String())
+	if err != nil {
+		if resp != nil {
+			defer resp.Body.Close()
+		}
+		return nil, err
+	}
+	if resp.StatusCode == 200 || resp.StatusCode == 201 {
+		return resp.Body, nil
+	} else {
+		return nil, fmt.Errorf("Failed to retrieve %s, response code %d", url.String(), resp.StatusCode)
+	}
+}
+
+func readerFromFileUrl(url *url.URL) (io.ReadCloser, error) {
+	return os.Open(url.Path)
+}
 
 // Determine whether a file exists in a container.
 func FileExistsInContainer(dockerClient *docker.Client, cId string, path string) bool {
@@ -84,6 +121,40 @@ func tarDirectory(dir string) (*os.File, error) {
 	return fw, nil
 }
 
+func downloadFile(url *url.URL, targetFile string, verbose bool) error {
+	sr := schemeReaders[url.Scheme]
+
+	reader, err := sr(url)
+
+	defer reader.Close()
+
+	if err != nil {
+		log.Printf("ERROR: Reading error while downloading %s (%s)\n", url.String(), err)
+		return err
+	}
+
+	out, err := os.Create(targetFile)
+	defer out.Close()
+
+	if err != nil {
+		defer os.Remove(targetFile)
+		log.Printf("ERROR: Unable to create target file %s (%s)\n", targetFile, err)
+		return err
+	}
+
+	_, err = io.Copy(out, reader)
+
+	if err != nil {
+		defer os.Remove(targetFile)
+		log.Printf("Skipping file %s due to error copying from source: %s\n", targetFile, err)
+	}
+
+	if verbose {
+		log.Printf("Downloaded '%s'\n", url.String())
+	}
+	return nil
+}
+
 func copy(sourcePath string, targetPath string) error {
 	info, err := os.Stat(sourcePath)
 	if err != nil {
@@ -111,4 +182,42 @@ func imageHasEntryPoint(image *docker.Image) bool {
 	}
 
 	return found
+}
+
+func executeCallback(callbackUrl string, result *BuildResult) {
+	buf := new(bytes.Buffer)
+	writer := bufio.NewWriter(buf)
+	for _, message := range result.Messages {
+		fmt.Fprintln(writer, message)
+	}
+	writer.Flush()
+
+	d := map[string]interface{}{
+		"payload": buf.String(),
+		"success": result.Success,
+	}
+
+	jsonBuffer := new(bytes.Buffer)
+	writer = bufio.NewWriter(jsonBuffer)
+	jsonWriter := json.NewEncoder(writer)
+	jsonWriter.Encode(d)
+	writer.Flush()
+
+	var resp *http.Response
+	var err error
+
+	for retries := 0; retries < 3; retries++ {
+		resp, err = http.Post(callbackUrl, "application/json", jsonBuffer)
+		if err != nil {
+			errorMessage := fmt.Sprintf("Unable to invoke callback: %s", err.Error())
+			result.Messages = append(result.Messages, errorMessage)
+		}
+		if resp != nil {
+			if resp.StatusCode >= 300 {
+				errorMessage := fmt.Sprintf("Callback returned with error code: %d", resp.StatusCode)
+				result.Messages = append(result.Messages, errorMessage)
+			}
+			break
+		}
+	}
 }
